fixed endpoints to send actual user data with game state

+++ src/main/java/com/aquino/TexasWebService/texas/TexasGame.java
@@ -387,7 +387,7 @@ public class TexasGame implements CardGame {
     }
     
     @JsonIgnore
-    public User getUser(int userId) {
+    public User getUser(long userId) {
         return userList.get(userId);
     }
